Extract useActionError hook from App.tsx
Moves the action error state, timeout ref, clearActionError/showActionError
callbacks, and cleanup effect into a dedicated hook at
src/tui/hooks/useActionError.ts. No logic changes.
